panda_backend/requirements.txt
asgiref==3.4.0
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
et-xmlfile==1.1.0
openpyxl==3.0.7
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
xlrd==2.0.1
